+++ rust/utils.rs
@@ -1,6 +1,6 @@
 /// O(v) where v is the number of voters
 // benchmarks show multi-threading with chunks then reduce is slower
-pub fn count_freqs(ballots: &[usize], n_parties: usize) -> Vec<u32> {
+pub fn count_freqs(ballots: &[usize], n_parties: usize) -> Vec<usize> {
     let mut counts = vec![0; n_parties];
     for ballot in ballots {
         counts[*ballot] += 1;
